Refactor TimeSeriesChart and related logic
Abstracted and modularized the TimeSeriesChart component by moving types, hooks, and utilities into separate files to improve maintainability and readability. Reorganized D3 rendering logic into a custom `useTimeSeriesD3` hook, while interaction and styling logic were separated for better reusability and testing.
