+++ frontend/src/api/hooks/useUser.ts
@@ -1,18 +1,18 @@
-import { useQuery, useMutation } from '@apollo/client';
-import { GET_USER } from '../queries/user';
-import { UPDATE_USER } from '../mutations/updateUser';
+import { useQuery, useMutation } from "@apollo/client";
+import { GET_USER } from "../queries/user";
+import { UPDATE_USER } from "../mutations/updateUser";
 import {
   UserResponse,
   UpdateUserMutationVariables,
   UpdateUserResponse,
-} from '@/types/graphql';
-import { useErrorHandler } from '@/hooks/useErrorHandler';
+} from "@/types/graphql";
+import { useErrorHandler } from "@/hooks/useErrorHandler";
 
 export const useUser = () => {
   const { handleError, showSuccess } = useErrorHandler();
 
   const { data, loading, error } = useQuery<UserResponse>(GET_USER, {
-    onError: handleError
+    onError: handleError,
   });
 
   const [updateUser, { loading: updateLoading, error: updateError }] =
@@ -20,9 +20,9 @@ export const useUser = () => {
       // Refetch user data after mutation
       refetchQueries: [{ query: GET_USER }],
       onCompleted: () => {
-        showSuccess('User updated successfully');
+        showSuccess("User updated successfully");
       },
-      onError: handleError
+      onError: handleError,
     });
 
   return {
